+++ server/src/middleware/auth.ts
